Keep the correct arglists (and other metadata).
Editor can show the arglists stored in the function metadata.
Unfortunately, with `defn-specs` the arglists are always something
similar to ([& args14464]) instead of ([] [x] [x y] [x y & more]).

This fixes the issue by first defining the function normally and then
rebinding the var to the function with wrapped specs assertions.